Prior to this fix the Use-Agent was hard coded which means that there was no way to comply with OSM tiles usage policy which say that a valid User-Agent identifying application should be set.
Using this fix the default User-Agent is set based on Gradle project's name and version properties (i.e. JXMapViewer2/2.2-SNAPSHOT).
In addition to that it is possible to to set custom User-Agent in order to properly identify application that is actually using JXMapViewer2 library under the hood.

Also removed few print statements.